introduce StateListener
