+++ src/main/scala/io/smartdatalake/app/GlobalConfig.scala
@@ -20,11 +20,10 @@
 package io.smartdatalake.app
 
 import com.typesafe.config.Config
-import io.smartdatalake.util.misc.{MemoryUtils, SmartDataLakeLogger}
-import org.apache.spark.sql.SparkSession
 import configs.syntax._
 import io.smartdatalake.definitions.Environment
-import org.apache.spark.{ExecutorPlugin, SparkConf, SparkEnv}
+import io.smartdatalake.util.misc.{MemoryUtils, SmartDataLakeLogger}
+import org.apache.spark.sql.SparkSession
 
 /**
  * Global configuration options
@@ -35,7 +34,9 @@ import org.apache.spark.{ExecutorPlugin, SparkConf, SparkEnv}
  * @param memoryLogTimer enable periodic memory usage logging, see detailled configuration [[MemoryLogTimerConfig]]
  * @param shutdownHookLogger enable shutdown hook logger to trace shutdown cause
  */
-case class GlobalConfig( kryoClasses: Option[Seq[String]] = None, sparkOptions: Option[Map[String,String]] = None, enableHive: Boolean = true, memoryLogTimer: Option[MemoryLogTimerConfig] = None, shutdownHookLogger: Boolean = false )
+case class GlobalConfig( kryoClasses: Option[Seq[String]] = None, sparkOptions: Option[Map[String,String]] = None, enableHive: Boolean = true
+                       , memoryLogTimer: Option[MemoryLogTimerConfig] = None, shutdownHookLogger: Boolean = false
+                       , stateListeners: Option[Seq[StateListener]] = None)
 extends SmartDataLakeLogger {
 
   // start memory logger, else log memory once
@@ -65,52 +66,9 @@ extends SmartDataLakeLogger {
 }
 object GlobalConfig {
   private[smartdatalake] def from(config: Config): GlobalConfig = {
-    globalConfig = Some(config.get[Option[GlobalConfig]]("global").value.getOrElse(GlobalConfig()))
+    globalConfig = Some(config.get[Option[GlobalConfig]]("global").value.get)//.getOrElse(GlobalConfig()))
     globalConfig.get
   }
   // store global config to be used in MemoryLoggerExecutorPlugin
   var globalConfig: Option[GlobalConfig] = None
-}
-
-
-/**
- * Configuration for periodic memory usage logging
- *
- * @param intervalSec interval in seconds between memory usage logs
- * @param logLinuxMem enable logging linux memory
- * @param logLinuxCGroupMem enable logging details about linux cgroup memory
- * @param logBuffers enable logging details about different jvm buffers
- */
-case class MemoryLogTimerConfig(intervalSec: Int, logLinuxMem: Boolean = true, logLinuxCGroupMem: Boolean = false, logBuffers: Boolean = false ) {
-  import MemoryLogTimerConfig._
-  private[smartdatalake] def startTimer(): Unit = MemoryUtils.startMemoryLogger(intervalSec, logLinuxMem, logLinuxCGroupMem, logBuffers)
-  private[smartdatalake] def getAsMap: Map[String, String] = Map(INTERVAL_SEC_OPTION -> intervalSec.toString, LOG_LINUX_MEM_OPTION -> logLinuxMem.toString, LOG_LINUX_CGROUP_MEM_OPTION -> logLinuxCGroupMem.toString, LOG_BUFFERS_OPTION -> logBuffers.toString)
-}
-private[smartdatalake] object MemoryLogTimerConfig{
-  def from(conf: SparkConf): MemoryLogTimerConfig = MemoryLogTimerConfig(conf.get(INTERVAL_SEC_OPTION).toInt, conf.get(LOG_LINUX_MEM_OPTION).toBoolean, conf.get(LOG_LINUX_CGROUP_MEM_OPTION).toBoolean, conf.get(LOG_BUFFERS_OPTION).toBoolean)
-  private val INTERVAL_SEC_OPTION = "spark.smartdatalake.memoryLog.intervalSec"
-  private val LOG_LINUX_MEM_OPTION = "spark.smartdatalake.memoryLog.logLinuxMem"
-  private val LOG_LINUX_CGROUP_MEM_OPTION = "spark.smartdatalake.memoryLog.logLinuxCGroupMem"
-  private val LOG_BUFFERS_OPTION = "spark.smartdatalake.memoryLog.logBuffers"
-}
-
-/**
- * Executor plugin to start memory usage logging on executors
- */
-private[smartdatalake] class MemoryLoggerExecutorPlugin extends ExecutorPlugin with SmartDataLakeLogger {
-  override def init(): Unit = {
-    // config can only be transferred to ExecutorPlugin by spark-options
-    val sparkConf = SparkEnv.get.conf
-    logger.debug("sparkConf: " + sparkConf.getAll.map{ case (k,v) => s"$k=$v"}.mkString(" "))
-    try {
-      val memoryLogConfig = MemoryLogTimerConfig.from(sparkConf)
-      memoryLogConfig.startTimer()
-      logger.info("MemoryLoggerExecutorPlugin successfully initialized")
-    } catch {
-      case e: Exception => logger.error(s"cannot initialize MemoryLoggerExecutorPlugin: ${e.getClass.getSimpleName} - ${e.getMessage}")
-    }
-  }
-  override def shutdown(): Unit = {
-    MemoryUtils.stopMemoryLogger()
-  }
 }
